automation/requirements.txt
playwright~=1.47.0
python-dotenv~=1.0.1
dnspython~=2.7.0
idna~=3.10
pip~=24.2
pymongo~=4.10.1
pyee~=12.0.0
typing_extensions~=4.12.2
pytest~=8.3.3
pluggy~=1.5.0
iniconfig~=2.0.0
certifi~=2024.8.30
requests~=2.32.3
python-slugify~=8.0.4
urllib3~=2.2.3
greenlet~=3.0.3
charset-normalizer~=3.4.0